requirements.txt
Django==4.0.2
asgiref==3.5.0
django-filter==21.1
djangorestframework==3.13.1
pytz==2021.3
sqlparse==0.4.2
psycopg2>=2.8
